cleaning step and moved stuff to resolver project
