+++ src/Monai.Deploy.WorkflowManager.ConditionsResolver.Tests/Resolver/ConditionalGroupTests.cs
@@ -1,11 +1,7 @@
-﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
-// SPDX-License-Identifier: Apache License 2.0
-
-using System;
-using Monai.Deploy.WorkflowManager.Contracts.Models;
+using Monai.Deploy.WorkflowManager.ConditionsResolver.Resolver;
 using Xunit;
 
-namespace Monai.Deploy.WorkflowManager.WorkflowExecuter.Tests.Common
+namespace Monai.Deploy.WorkflowManager.ConditionsResolver.Tests.Resolver
 {
     public class ConditionalGroupTests
     {
@@ -64,29 +60,5 @@ namespace Monai.Deploy.WorkflowManager.WorkflowExecuter.Tests.Common
         //    var result = conditional.Evaluate();
         //    Assert.Equal(expected, result);
         //}
-
-        [Theory]
-        [InlineData("{{context.dicom.tags[('0010','0040')]}}", "F", "{{context.dicom.tags[('0010','0040')]}} == 'F'")]
-        [InlineData("{{context.executions.body_part_identifier.result.body_part}}", "leg", "{{context.executions.body_part_identifier.result.body_part}} == 'leg'")]
-        public void Conditional_CreatesAndEvaluates(string expectedLeftParam, string expectedRightParam, string input)
-        {
-            var conditional = Conditional.Create(input);
-            var leftParameter = conditional.LeftParameter;
-            var rightParameter = conditional.RightParameter;
-
-            Assert.Equal(expectedLeftParam, leftParameter);
-            Assert.Equal(expectedRightParam, rightParameter);
-        }
-
-        [Theory]
-        [InlineData("AND {{context.dicom.tags[('0010','0040')]}} == 'F'", "No left hand parameter at index: 0")]
-        [InlineData(" AND {{context.dicom.tags[('0010','0040')]}} == 'F'", "No left hand parameter at index: 0")]
-        //[InlineData("OR {{context.dicom.tags[('0010','0040')]}} == 'F'", "No left hand parameter at index: 0")]
-        public void Conditional_CreatesAndEvaluates_ErrorsOnInvalidInput(string input, string expectedErrorMessage)
-        {
-            var exception = Assert.Throws<ArgumentException>(() => Conditional.Create(input));
-
-            Assert.Equal(expectedErrorMessage, exception.Message);
-        }
     }
 }
