+++ src/Monai.Deploy.WorkflowManager.ConditionsResolver/Resovler/Keyword.cs
@@ -1,0 +1,12 @@
+﻿namespace Monai.Deploy.WorkflowManager.ConditionsResolver.Resolver
+{
+    /// <summary>
+    /// Group Keywords or Operators
+    /// </summary>
+    public enum Keyword
+    {
+        SINGULAR,
+        AND,
+        OR,
+    }
+}
