+++ src/Monai.Deploy.WorkflowManager.sln
@@ -35,7 +35,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManage
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.Database.Tests", "Database.Tests\Monai.Deploy.WorkflowManager.Database.Tests.csproj", "{F1F553CD-373E-4D82-9A92-6D138C4F0A5A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monai.Deploy.WorkflowManager.Common.Tests", "..\tests\Common.Tests\Monai.Deploy.WorkflowManager.Common.Tests.csproj", "{3BFC16AF-8881-4B6A-B39C-495FB3484992}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.Common.Tests", "..\tests\Common.Tests\Monai.Deploy.WorkflowManager.Common.Tests.csproj", "{3BFC16AF-8881-4B6A-B39C-495FB3484992}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monai.Deploy.WorkflowManager.ConditionsResolver", "Monai.Deploy.WorkflowManager.ConditionsResolver\Monai.Deploy.WorkflowManager.ConditionsResolver.csproj", "{25C84A62-8994-4981-B765-C6FADF1BD991}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monai.Deploy.WorkflowManager.ConditionsResolver.Tests", "Monai.Deploy.WorkflowManager.ConditionsResolver.Tests\Monai.Deploy.WorkflowManager.ConditionsResolver.Tests.csproj", "{7EC65834-FB21-4325-B8B9-3578AF3C7FB1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -107,6 +111,14 @@ Global
 		{3BFC16AF-8881-4B6A-B39C-495FB3484992}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3BFC16AF-8881-4B6A-B39C-495FB3484992}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3BFC16AF-8881-4B6A-B39C-495FB3484992}.Release|Any CPU.Build.0 = Release|Any CPU
+		{25C84A62-8994-4981-B765-C6FADF1BD991}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{25C84A62-8994-4981-B765-C6FADF1BD991}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{25C84A62-8994-4981-B765-C6FADF1BD991}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{25C84A62-8994-4981-B765-C6FADF1BD991}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7EC65834-FB21-4325-B8B9-3578AF3C7FB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7EC65834-FB21-4325-B8B9-3578AF3C7FB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7EC65834-FB21-4325-B8B9-3578AF3C7FB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7EC65834-FB21-4325-B8B9-3578AF3C7FB1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -119,6 +131,7 @@ Global
 		{6E095A30-EDF0-4940-935D-A8B64C19DE22} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 		{F1F553CD-373E-4D82-9A92-6D138C4F0A5A} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 		{3BFC16AF-8881-4B6A-B39C-495FB3484992} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
+		{7EC65834-FB21-4325-B8B9-3578AF3C7FB1} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DC0D56C8-D8CB-45CE-B528-F3DCF86D63ED}
